fixing performance issues by improving schedule screen architecture
